nova migracao

+++ api/evaluation/models.py
@@ -4,14 +4,14 @@ from django.db import models
 
 # Create your models here.
 class MRR(models.Model):
-    value = models.IntegerField(unique=False)
+    value = models.FloatField(unique=False)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
 class MAP(models.Model):
-    value = models.IntegerField(unique=False)
+    value = models.FloatField(unique=False)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
 class NDCG(models.Model):
-    value = models.IntegerField(unique=False)
+    value = models.FloatField(unique=False)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
